\name{computeDayLength}
\alias{computeDayLength}
\title{computeDayLength}
\description{Compute the Day-length in hours for given time and coordinates}
\usage{computeDayLength(timestamp, latDeg, ...)}
\arguments{
  \item{timestamp}{POSIXt vector}
  \item{latDeg}{Latitude in (decimal) degrees}
  \item{\dots}{further arguments to \code{\link{computeDayLengthDoy}}}
}

\value{result of \code{\link{computeDayLengthDoy}}}

\author{Thomas Wutzler}





