\name{getFractionalHours}
\alias{getFractionalHours}
\title{getFractionalHours}
\description{get the time difference to previous midnight in fractional hours}
\usage{getFractionalHours(timestamp)}
\arguments{
  \item{timestamp}{POSIXt vector}
}

\value{numeric vector of fractional hours}

\author{Thomas Wutzler}





