\name{getHoursAheadOfUTC}
\alias{getHoursAheadOfUTC}
\title{getHoursAheadOfUTC}
\description{get the time difference to UTC in hours}
\usage{getHoursAheadOfUTC(timestamp)}
\arguments{
  \item{timestamp}{POSIXt vector}
}

\value{integer vector of how many hours noon of timestamp is ahead of
noon in UTC}

\author{Thomas Wutzler}





