\name{getSolarTimeHour}
\alias{getSolarTimeHour}
\title{getSolarTimeHour}
\description{Get the fractional hour of solar time}
\usage{getSolarTimeHour(timestamp, longDeg)}
\arguments{
  \item{timestamp}{POSIXt vector in local time}
  \item{longDeg}{Longitude in (decimal) degrees}
}

\value{fractional hour corrected by difference to local time}

\author{Thomas Wutzler}





