\name{setLocalTimeZone}
\alias{setLocalTimeZone}
\title{setLocalTimeZone}
\description{modify tzone attribute of timestamp to 'GMT+x' for local to given longitude}
\usage{setLocalTimeZone(timestamp, longDeg)}
\arguments{
  \item{timestamp}{POSIXct}
  \item{longDeg}{Longitude in (decimal) degrees}
}

\value{\code{timestamp} with modified tzone attribute}

\author{Thomas Wutzler}





