\name{computeSunPosition}
\alias{computeSunPosition}
\title{computeSunPosition}
\description{Calculate the position of the sun}
\usage{computeSunPosition(timestamp, latDeg, longDeg)}
\arguments{
  \item{timestamp}{POSIXct}
  \item{latDeg}{Latitude in (decimal) degrees}
  \item{longDeg}{Longitude in (decimal) degrees}
}

\value{as returned by \code{\link{computeSunPositionDoyHour}}}

\author{Thomas Wutzler}





